# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.2.0
  description: Unified API for GambiaInfluencers and Our Influencers platforms
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /oi/users:
    get:
      operationId: oiListUsers
      tags: [users]
      summary: List all users (admin only)
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OiUser"
    post:
      operationId: oiCreateUser
      tags: [users]
      summary: Create admin or client user (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OiCreateUserBody"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OiUser"
  /oi/users/{id}:
    patch:
      operationId: oiUpdateUser
      tags: [users]
      summary: Update user (admin only)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OiUpdateUserBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OiUser"
    delete:
      operationId: oiDeleteUser
      tags: [users]
      summary: Delete user (admin only)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── Our Influencers: Campaigns ──────────────────────────────────────────────
...
components:
  schemas:
...
    OiUser:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        fullName:
          type: string
        role:
          type: string
          enum: [admin, client]
        companyName:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, email, fullName, role, companyName, phone, createdAt]
    OiCreateUserBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
        fullName:
          type: string
        role:
          type: string
          enum: [admin, client]
        companyName:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
      required: [email, password, fullName, role]
    OiUpdateUserBody:
      type: object
      properties:
        email:
          type: string
        fullName:
          type: string
        companyName:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        password:
          type: ["string", "null"]
...