# lib/api-zod/src/generated/types/oiUpdateUserBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Unified API for GambiaInfluencers and Our Influencers platforms
 * OpenAPI spec version: 0.2.0
 */

export interface OiUpdateUserBody {
  email?: string;
  fullName?: string;
  /** @nullable */
  companyName?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  password?: string | null;
}